lambda/requirement.txt
boto3
Pillow


